Create checkarrayvalue.java
Write a Java program to test if 10 appears as either the first or last element of an array of integers. The length of the array must be greater than or equal to 2. Go to the editor
Sample OutputWrite a Java program to test if 10 appears as either the first or last element of an array of integers. The length of the array must be greater than or equal to 2. Go to the editor
Sample Output